fix: fix modules
