+++ js/main.js
@@ -1,4 +1,4 @@
-import {createHostelsData} from './create-hotel.js';
+import {createHostelsData} from './data.js';
 
 // eslint-disable-next-line no-console
 console.log(createHostelsData());
